Add support for using retworkx as input and output from Matching
This commit adds support for using retworkx [1] as a graph library input
and output in addition to NetworkX. This enables users to go from a
retworkx PyGraph object to a PyMatching Matching object.

The retworkx library is a higher performance general purpose graph
library written in Rust that was inspired by NetworkX. While similar to
NetworkX it is not a drop in replacement (see [2] for details of the
differences). retworkx was originally written to accelerate the graph data
structures and algorithms used internally by Qiskit and has since grown to
be a high performance general graph library that is suitable for any
application. Since it's adoption by Qiskit the library has gained some
adoption in several QEC libraries such as flamingpy [3] and qtcodes [4],
along with Qiskit's topological codes module in qiskit-ignis (which was
recently spun off into a separate repository). [5] Being able to leverage
retworkx graphs directly will simplify potentially using PyMatching in any
of these packages.

For this commit I opted to add retworkx to the requirements list, but it
could easily be made an optional dependency if that would be preferrable
so that two graph libraries aren't required to install PyMatching.

[1] https://github.com/Qiskit/retworkx
[2] https://qiskit.org/documentation/retworkx/networkx.html
[3] https://github.com/XanaduAI/flamingpy
[4] https://github.com/yaleqc/qtcodes
[5] https://github.com/NCCR-SPIN/topological_codes